+++ flashinfer/utils.py
@@ -21,7 +21,6 @@ from typing import Callable, Dict, Iterable, Optional, Sequence, Tuple, Union
 
 import torch
 import torch.version
-from einops import rearrange, reduce, repeat
 from torch.torch_version import TorchVersion
 from torch.torch_version import __version__ as torch_version
 
@@ -490,183 +489,3 @@ def ceil_div(x: int, y: int) -> int:
         The result of the ceiling division.
     """
     return (x + y - 1) // y
-
-
-def _ceil_to_ue8m0(x: torch.Tensor):
-    """imported from DeepGEMM"""
-    assert x.view(-1).amax().item() > 0
-    return torch.pow(2.0, torch.ceil(torch.log2(x.abs())))
-
-
-def per_token_cast_to_fp8(x: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
-    """imported from DeepGEMM"""
-    assert x.dim() == 2 and x.size(1) % 128 == 0
-    m, n = x.shape
-    x_view = x.view(m, -1, 128)
-    x_amax = x_view.abs().float().amax(dim=2).view(m, -1).clamp(1e-4)
-    sf = _ceil_to_ue8m0(x_amax / 448.0)
-    return (x_view * (1.0 / sf.unsqueeze(2))).to(torch.float8_e4m3fn).view(m, n), sf
-
-
-def per_block_cast_to_fp8(x: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
-    """imported from DeepGEMM"""
-    assert x.dim() == 2
-    m, n = x.shape
-    x_padded = torch.zeros(
-        (round_up(m, 128), round_up(n, 128)), dtype=x.dtype, device=x.device
-    )
-    x_padded[:m, :n] = x
-    x_view = x_padded.view(-1, 128, x_padded.size(1) // 128, 128)
-    x_amax = x_view.abs().float().amax(dim=(1, 3), keepdim=True).clamp(1e-4)
-    sf = _ceil_to_ue8m0(x_amax / 448.0)
-    x_scaled = (x_view * (1.0 / sf)).to(torch.float8_e4m3fn)
-    return x_scaled.view_as(x_padded)[:m, :n].contiguous(), sf.view(
-        x_view.size(0), x_view.size(2)
-    )
-
-
-def quantize_fp8(x, scale_shape, tile_shape, scale_major_mode):
-    """
-    Quantizes a 2D or 3D tensor to FP8.
-
-    Args:
-        x (torch.Tensor): The 2D or 3D input tensor.
-        scale_shape (tuple): The shape of the scale tensor.
-        tile_shape (tuple): The shape of the tiles.
-        scale_major_mode (str): The tiling order, "K" for row-major like,
-                                or another value for column-major like.
-
-    Returns:
-        tuple: A tuple containing the quantized FP8 tensor and the
-               calculated float32 scales.
-    """
-    # 1. Assertions and Initial Setup
-    ndim = x.ndim
-    assert ndim in [2, 3], f"x.ndim must be 2 or 3, but got {ndim}"
-    assert ndim == len(scale_shape) == len(tile_shape)
-
-    fp8_info = torch.finfo(torch.float8_e4m3fn)
-    fp8_amax = torch.tensor(fp8_info.max, device=x.device, dtype=torch.float32)
-
-    # 2. Tiling and Scale Calculation
-    if ndim == 2:
-        s0, s1 = scale_shape
-        t0, t1 = tile_shape
-        if scale_major_mode == "K":
-            # Tile x and find the max absolute value in each tile
-            x_tiled = rearrange(x, "(s0 t0) (s1 t1) -> s0 s1 t0 t1", s0=s0, s1=s1)
-            abs_max = reduce(x_tiled.abs(), "s0 s1 t0 t1 -> s0 s1", "max").clamp(1e-4)
-            x_scale = abs_max / fp8_amax
-            x_scale = torch.pow(2.0, torch.ceil(torch.log2(x_scale.abs())))
-
-            # Broadcast scales back to the original tensor shape
-            scales_repeated = repeat(x_scale, "s0 s1 -> (s0 t0) (s1 t1)", t0=t0, t1=t1)
-        else:
-            # Handle column-major tiling
-            x_tiled = rearrange(x, "(s1 t0) (s0 t1) -> s0 s1 t0 t1", s0=s0, s1=s1)
-            abs_max = reduce(x_tiled.abs(), "s0 s1 t0 t1 -> s0 s1", "max").clamp(1e-4)
-            x_scale = abs_max / fp8_amax
-            x_scale = torch.pow(2.0, torch.ceil(torch.log2(x_scale.abs())))
-
-            # Permute scale axes before repeating to match layout
-            scales_permuted = rearrange(x_scale, "s0 s1 -> s1 s0")
-            scales_repeated = repeat(
-                scales_permuted, "s1 s0 -> (s1 t0) (s0 t1)", t0=t0, t1=t1
-            )
-
-    elif ndim == 3:
-        s0, s1, s2 = scale_shape
-        t0, t1, t2 = tile_shape
-        if scale_major_mode == "K":
-            # Tile x and find the max absolute value in each tile
-            x_tiled = rearrange(
-                x, "(s0 t0) (s1 t1) (s2 t2) -> s0 s1 s2 t0 t1 t2", s0=s0, s1=s1, s2=s2
-            )
-            abs_max = reduce(
-                x_tiled.abs(), "s0 s1 s2 t0 t1 t2 -> s0 s1 s2", "max"
-            ).clamp(1e-4)
-            x_scale = abs_max / fp8_amax
-            x_scale = torch.pow(2.0, torch.ceil(torch.log2(x_scale.abs())))
-
-            # Broadcast scales back to the original tensor shape
-            scales_repeated = repeat(
-                x_scale, "s0 s1 s2 -> (s0 t0) (s1 t1) (s2 t2)", t0=t0, t1=t1, t2=t2
-            )
-        else:
-            # Handle layout where the last two axes are swapped
-            x_tiled = rearrange(
-                x, "(s0 t0) (s2 t1) (s1 t2) -> s0 s1 s2 t0 t1 t2", s0=s0, s1=s1, s2=s2
-            )
-            abs_max = reduce(
-                x_tiled.abs(), "s0 s1 s2 t0 t1 t2 -> s0 s1 s2", "max"
-            ).clamp(1e-4)
-            x_scale = abs_max / fp8_amax
-            x_scale = torch.pow(2.0, torch.ceil(torch.log2(x_scale.abs())))
-
-            # Permute scale axes before repeating to match layout
-            scales_permuted = rearrange(x_scale, "s0 s1 s2 -> s0 s2 s1")
-            scales_repeated = repeat(
-                scales_permuted,
-                "s0 s2 s1 -> (s0 t0) (s2 t1) (s1 t2)",
-                t0=t0,
-                t1=t1,
-                t2=t2,
-            )
-
-    # 3. Final Quantization
-    # Divide the original tensor by the broadcasted scales
-    x_fp32 = x / (scales_repeated + 1e-8)
-
-    # Convert the result to the target FP8 format
-    x_fp8 = x_fp32.to(torch.float8_e4m3fn)
-
-    return x_fp8, x_scale
-
-
-def dequantize_fp8(x, x_scale, scale_major_mode):
-    """
-    Quantizes a 2D or 3D tensor to FP8.
-
-    Args:
-        x (torch.Tensor): The 2D or 3D input tensor.
-        scale_shape (tuple): The shape of the scale tensor.
-        tile_shape (tuple): The shape of the tiles.
-        scale_major_mode (str): The tiling order, "K" for row-major like,
-                                or another value for column-major like.
-
-    Returns:
-        tuple: A tuple containing the quantized FP8 tensor and the
-               calculated float32 scales.
-    """
-    # 1. Assertions and Initial Setup
-    ndim = x.ndim
-    assert ndim in [2, 3], f"x.ndim must be 2 or 3, but got {ndim}"
-    assert ndim == len(x_scale.shape)
-
-    # 2. Tiling and Scale Calculation
-    if ndim == 2:
-        s0, s1 = x_scale.shape
-        x = rearrange(
-            x.to(torch.float32), "(s0 t0) (s1 t1) -> s0 s1 t0 t1", s0=s0, s1=s1
-        )
-        if scale_major_mode == "K":
-            x_scale = rearrange(x_scale, "s0 s1 -> s0 s1 1 1")
-        else:
-            x_scale = rearrange(x_scale, "s0 s1 -> s1 s0 1 1")
-        out = rearrange(x * x_scale, "s0 s1 t0 t1 -> (s0 t0) (s1 t1)")
-
-    elif ndim == 3:
-        s0, s1, s2 = x_scale.shape
-        x = rearrange(
-            x.to(torch.float32),
-            "(s0 t0) (s1 t1) (s2 t2)-> s0 s1 s2 t0 t1 t2",
-            s0=s0,
-            s1=s1,
-            s2=s2,
-        )
-        if scale_major_mode == "K":
-            x_scale = rearrange(x_scale, "s0 s1 s2 -> s0 s1 s2 1 1 1")
-        else:
-            x_scale = rearrange(x_scale, "s0 s1 s2 -> s0 s2 s1 1 1 1")
-        out = rearrange(x * x_scale, "s0 s1 s2 t0 t1 t2 -> (s0 t0) (s1 t1) (s2 t2)")
-    return out
